03-feature-forge/pyproject.toml
[project]
name = "feature-forge"
version = "0.1.0"
description = "FeatureForge - ML Feature Store and Drift Detection System built on Snowflake"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "FeatureForge Team" }]

dependencies = [
    "snowflake-connector-python>=3.6.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "apache-airflow>=2.8.0",
    "boto3>=1.34.0",
    "great-expectations>=0.18.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pandas-stubs>=2.1.0",
    "boto3-stubs[sagemaker]>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "snowflake.connector.*",
    "great_expectations.*",
    "airflow.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
